feat: add BalancesTable to AddressPage
