+++ package.json
@@ -17,7 +17,7 @@
     "@types/react-dom": "^17.0.11",
     "@types/styled-components": "^5.1.15",
     "bignumber.js": "^9.0.1",
-    "fuels": "^0.1.1",
+    "fuels": "^0.0.0-master-9a71ed8",
     "graphql": "^15.7.2",
     "qrcode.react": "^1.0.1",
     "react": "^17.0.2",
